+++ changelog.md
@@ -13,10 +13,12 @@
   - Implement Facebook SSO
 - Settings:
   - Fix: Date of Birth cleared when switching to an account without a set DOB
-- Icons:
+- Main menu:
   - Brand icons migrated to Simple Icons (No longer the same style either)
   - Remove Linkedin link
-- Gameplay:
+  - `Continue Game` menu item text changed to include level number:
+    `Continue Level ${lastLevel}`
+- Gameplay UI:
   - Vertically center game canvas
 
 # 0.8.12 Beta - 20/03/2026 23:30 UTC+1